# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Office Attendance Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DayState = typeof DayState[keyof typeof DayState];


export const DayState = {
  present: 'present',
  company_leave: 'company_leave',
  personal_leave: 'personal_leave',
  planned: 'planned',
  remote: 'remote',
} as const;

export interface AttendanceDay {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  state: DayState;
  /** @nullable */
  note?: string | null;
  createdAt?: string;
  updatedAt?: string;
}

export interface AttendanceInput {
  /** Date in YYYY-MM-DD format */
  date: string;
  state: DayState;
  note?: string;
}

export interface DeleteResult {
  success: boolean;
}

export type BulkAttendanceBodyRecordsItemAction = typeof BulkAttendanceBodyRecordsItemAction[keyof typeof BulkAttendanceBodyRecordsItemAction];


export const BulkAttendanceBodyRecordsItemAction = {
  upsert: 'upsert',
  delete: 'delete',
} as const;

export type BulkAttendanceBodyRecordsItem = {
  date: string;
  action: BulkAttendanceBodyRecordsItemAction;
  state?: DayState;
  note?: string;
};

export interface BulkAttendanceBody {
  records: BulkAttendanceBodyRecordsItem[];
}

export interface BulkResult {
  updated: number;
  deleted: number;
}

export interface QuarterSummary {
  /** Quarter number (1-4) */
  quarter: number;
  /** e.g. "Q1 Jan - Mar" */
  label: string;
  months: number[];
  totalWeeks: number;
  /** Weeks with zero office presence (past weeks only) */
  wfhWeeks: number;
  presentDays: number;
  plannedDays: number;
  personalLeaveDays: number;
  companyLeaveDays: number;
  remoteDays: number;
  totalWorkdays: number;
}

export interface YearSummary {
  year: number;
  totalPresent: number;
  totalPlanned: number;
  totalPersonalLeave: number;
  totalCompanyLeave: number;
  totalRemote: number;
  totalWorkdays: number;
}

export type ListAttendanceParams = {
year?: number;
month?: number;
};

export type GetQuarterlySummaryParams = {
year?: number;
};

export type GetYearSummaryParams = {
year?: number;
};