# lib/api-zod/src/generated/types/bulkResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Office Attendance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface BulkResult {
  updated: number;
  deleted: number;
}