# lib/api-zod/src/generated/types/quarterSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Office Attendance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface QuarterSummary {
  /** Quarter number (1-4) */
  quarter: number;
  /** e.g. "Q1 Jan - Mar" */
  label: string;
  months: number[];
  totalWeeks: number;
  /** Weeks with zero office presence (past weeks only) */
  wfhWeeks: number;
  presentDays: number;
  plannedDays: number;
  personalLeaveDays: number;
  companyLeaveDays: number;
  remoteDays: number;
  totalWorkdays: number;
}